+++ runner/monitor/CHANGELOG.md
@@ -5,6 +5,7 @@
 `androidx.test:monitor:{version}` is released.
 
 **Bug Fixes**
+* Activities can be started from any thread.
 
 **New Features**
 
@@ -14,6 +15,7 @@
 
 * Update to minSdkVersion 21
 * Make ReflectionException a RuntimeException
+* Hid IntentStubberRegistry.getInstance()
 
 **Breaking API Changes**